src/proxy/requirements.txt
requests==2.24.0
pyJWT==2.0.1
Flask==1.1.1
